% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtfn-package.R
\docType{package}
\name{wtfn-package}
\alias{wtfn-package}
\title{wtfn: Where's This Function?}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rossellhayes/wtfn}
  \item Report bugs at \url{https://github.com/rossellhayes/wtfn/issues}
}

}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
