% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtfn.R
\name{wtfn}
\alias{wtfn}
\title{Where's this function?}
\usage{
wtfn(fun, verbose = TRUE)
}
\arguments{
\item{fun}{A function, as either a character string or a symbol}

\item{verbose}{A \link{logical}.
If \code{verbose} is \code{\link{TRUE}}, \code{wtfn()} prints all information about \code{fun}.
If \code{verbose} is \code{\link{FALSE}}, \code{wtfn()} prints a single line indicating if you
can use \code{fun} in your package given its current dependencies.
Defaults to \code{\link{TRUE}}.}
}
\value{
Invisibly returns \code{\link{TRUE}} if you can use \code{fun} in your package
given your current dependencies and \code{\link{FALSE}} if you cannot.
}
\description{
Prints information about the package containing a function and whether
you can use that function given the current dependencies of your package.
}
\examples{
wtfn("paste")
wtfn("wtfn::wtfn")
}
