% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caniuse-package.R
\docType{package}
\name{caniuse-package}
\alias{caniuse}
\alias{caniuse-package}
\title{caniuse: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alexander Rossell Hayes \email{alexander@rossellhayes.com} (\href{https://orcid.org/0000-0001-9412-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
